Add files via upload
According to the Data Set has more than 5 million records are being stolen on a daily basis, a concerning statistic that shows - fraud is still very common both for Card-Present and Card-not Present type of payments. the Digital payments are evolving, but so are cyber criminals. we experiment with the creditcard fraud dataset to explore the machine learning algorithms and build an optimum model to predict the fraud.